se logra end to end del findall de review
