+++ src/main/java/ar/edu/unq/desapp/grupoI/backenddesappapi/WebSecurityConfiguration.java
@@ -24,6 +24,7 @@ public class WebSecurityConfiguration extends WebSecurityConfigurerAdapter {
                 .antMatchers(HttpMethod.POST, "/newReview").permitAll()
                 .antMatchers(HttpMethod.POST, "/newuser/*").permitAll()
                 .antMatchers(HttpMethod.GET, "/hello").permitAll()
+                .antMatchers(HttpMethod.GET, "/reviews").permitAll()
                 .antMatchers(HttpMethod.GET, "/review/reviews").permitAll()
                 .antMatchers(HttpMethod.GET, "/api/hello").permitAll()
                 .anyRequest().authenticated();
