+++ src/main/java/ar/edu/unq/desapp/grupoI/backenddesappapi/dto/ReviewDTO.java
@@ -1,0 +1,19 @@
+package ar.edu.unq.desapp.grupoi.backenddesappapi.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDate;
+
+@Getter
+@Setter
+public class ReviewDTO {
+    public String tittle_tconst;
+    public String resume;
+    public String extendedText;
+    public float rating;
+    public LocalDate date;
+    public String language;
+    public Long userId;
+    public boolean spoiler_Alert;
+}
