+++ src/main/java/ar/edu/unq/desapp/grupoI/backenddesappapi/repositories/review/ReviewRepositoryImpl.java
@@ -3,22 +3,26 @@ package ar.edu.unq.desapp.grupoi.backenddesappapi.repositories.review;
 
 import ar.edu.unq.desapp.grupoi.backenddesappapi.exceptions.ReviewHasBeenAddedException;
 import ar.edu.unq.desapp.grupoi.backenddesappapi.model.reviews.Review;
+import org.springframework.stereotype.Repository;
 
+import javax.persistence.EntityManager;
+import javax.persistence.TypedQuery;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
+@Repository
 public class ReviewRepositoryImpl {
 
     List<Review> reviews;
+    private EntityManager em;
 
+    public ReviewRepositoryImpl(EntityManager em) {
+        this.em = em;
+    }
     public ReviewRepositoryImpl() {
         this.reviews = new ArrayList<>();
-    }
-
-
-    public List<Review> findAll() {
-        return this.reviews;
     }
 
     public List<Review> getReviews(String tittle_tconst) {
